requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
openpyxl>=3.1.0
grad-cam>=1.4.8
umap-learn>=0.5.0
tqdm>=4.65.0
optuna>=3.5.0
imagehash>=4.3.1
